feat: 공용 버튼 생성 및 파이 수정
